backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pandas==2.3.3
openpyxl==3.1.5
xlsxwriter==3.2.9
openai==2.8.1
python-dotenv==1.2.1
motor==3.6.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
